move src/qmlapplication.{cpp,h} to src/qml
